requirement.txt
pandas
numpy
seaborn
matplotlib
scikit-sklearn
catboost
xgboost
dill
#-e .